Refactor notification service into package
